requirements.txt
pyjwt==2.4.0
uuid==1.30
redis==4.3.4
pytest==7.0.1
requests==2.27.1